Update dependencies and workflows
This is a bit of an overhaul of the library with the main aim being
to include eslint/prettier as dependencies so we don't have to
directly include them in other projects and deal with updates there.

I also updated the CI scripts while I was here to match our flows in
other repos.